refactor: moved opengl related code into own package
