+++ imgui.ini
@@ -33,7 +33,7 @@ Size=339,290
 Collapsed=0
 
 [Window][window]
-ViewportPos=1021,257
+ViewportPos=1021,256
 ViewportId=0x8BE4F9DD
 Size=338,179
 Collapsed=0
