+++ src/main/java/io/nova/opengl/renderer/Batch.java
@@ -1,4 +1,4 @@
-package io.nova.renderer;
+package io.nova.opengl.renderer;
 
 import io.nova.components.Sprite;
 import io.nova.core.renderer.Camera;
