+++ src/main/java/io/nova/opengl/renderer/OpenGLTexture2d.java
@@ -1,4 +1,4 @@
-package io.nova.renderer;
+package io.nova.opengl.renderer;
 
 import io.nova.core.renderer.Texture2d;
 import org.lwjgl.BufferUtils;
